Starter Commit du 18-03-25

+++ src/app/music-card/music-card.component.ts
@@ -1,16 +1,18 @@
-import {Component, inject, Input, input, model} from '@angular/core';
+import {Component, inject, input} from '@angular/core';
 import {MatButtonModule} from '@angular/material/button';
 import {MatCardModule} from '@angular/material/card';
-import {Titre, User} from '../_model/model';
+import {Playlist, Titre} from '../_model/model';
 import {MatIcon} from '@angular/material/icon';
 import {MatMenu, MatMenuItem, MatMenuTrigger} from '@angular/material/menu';
-import {
-  MatDialog,
-} from '@angular/material/dialog';
+import {MatDialog} from '@angular/material/dialog';
 import {DialogAddPlaylistComponent} from '../dialog-add-playlist/dialog-add-playlist.component';
 import {AuthentificationService} from '../_services/authentification.service';
-import {DataServiceService} from '../_model/data-service.service';
+//import {DataServiceService} from '../_model/data-service.service';
 import {UserConnectedService} from '../_services/user-connected.service';
+import {Router} from '@angular/router';
+import {MatSnackBar} from '@angular/material/snack-bar';
+import {NotificationService} from '../_services/notification.service';
+//import {NotificationComponent} from '../notification/notification.component';
 
 @Component({
   selector: 'app-music-card',
@@ -22,9 +24,10 @@ import {UserConnectedService} from '../_services/user-connected.service';
 export class MusicCardComponent {
 
   //injection of authentication service, data service
-  //private authenticationService = inject(AuthentificationService);
+  private authenticationService = inject(AuthentificationService);
   //private dataService = inject(DataServiceService);
   private userConnectedService = inject(UserConnectedService);
+  private notificationService = inject(NotificationService);
 
   titre = input.required<Titre>();
   param = input.required<string>()
@@ -32,29 +35,49 @@ export class MusicCardComponent {
   playlists = this.userConnectedService.getPlaylistUser();
 
   readonly dialog = inject(MatDialog);
+  private router = inject(Router);
 
   openDialog(): void {
+    if(!this.authenticationService.getIsAuthenticated){
+      this.router.navigate(['/login']);
+      return;
+    }
     this.dialog.open(DialogAddPlaylistComponent, {
       width: '30%',
     });
   }
 
   sauvegarderTitre() {
+    if(!this.authenticationService.getIsAuthenticated){
+      this.router.navigate(['/login']);
+      return;
+    }
     this.userConnectedService.addSauvegarde(this.titre());
+    this.notificationService.show(`Titre ${this.titre().name} Sauvegardé`);
   }
 
   deleteToSauvegarde() {
-    this.userConnectedService.deleteSauvegardeUser(this.titre().id);
+    if (confirm("Voulez vous supprimer ce titre de la sauvegarde ?")) {
+      this.userConnectedService.deleteSauvegardeUser(this.titre().id);
+      this.notificationService.show(`Titre ${this.titre().name} supprimé de la sauvegarde`);
+    }
   }
 
   isInSauvegarde() : boolean {
-    return this.userConnectedService.isTitleSauvegarde(this.titre().id);
+    return this.userConnectedService.isTitleInSauvegarde(this.titre().id);
   }
 
 
   addToPlaylist(nom: string) {
-    this.userConnectedService.adTitleToPlaylist(nom, this.titre());
+    this.userConnectedService.addTitleToPlaylist(nom, this.titre());
+    this.notificationService.show(`Titre ${this.titre().name} ajouté à la playlist ${nom}`);
     console.log(this.userConnectedService.getPlaylistUser());
+  }
+
+
+  isInPlaylist(p: Playlist) {
+    const titre = p.titres.find(t => t.id == this.titre().id)
+    return !!titre;
   }
 }
